+++ routes/api/auth.js
@@ -2,6 +2,9 @@ const express = require("express");
 
 const router = express.Router();
 
-router.post("/register");
+const validateBodyRequest = require("../../middlewares/validate-body");
+const {registerSchema} = require("../../schemas")
+
+router.post("/register", validateBodyRequest(registerSchema));
 
 module.exports = router;
